requirements.txt
Django==4.2.4
djangorestframework==3.14.0
firebase_admin==6.2.0
fitz==0.0.1.dev2
matplotlib==3.7.2
numpy==1.25.2
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.8.0.76
paddleocr==2.7.0.2
paddleocr.egg==info
pdf2image==1.16.3
Pillow==10.0.0
Pillow==10.0.0
PyMuPDF==1.20.2
python-dotenv==1.0.0
Requests==2.31.0
setuptools==65.5.1
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.1
transformers==4.31.0
torchaudio==0.12.1+cu113
-f https://download.pytorch.org/whl/cu113/torch_stable.html
fitz
img2pdf
